Move jacobian computation out of SBSurfaceFluidCouplingManager.
